refactor: extract shared allowed-regions loader into a store
Both the completeness and detailed-data tables had their own near-identical
copy of the /api/user-regions fetch + allowedRegionIDs state. Extract it into
a single shared store (src/lib/stores/userRegions.ts) and have both components
import it.

- loadAllowedRegions() caches the in-flight promise so concurrent callers
  await the same fetch (a boolean guard would let a second caller return
  early against the default empty store and render nothing).
- Same fail-closed behavior: network / HTTP / parse / shape errors set the
  store to [] so a broken PMP connection never exposes all regions.
